requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
SQLAlchemy==2.0.40
cryptography>=42.0.0
Werkzeug>=3.1
Jinja2>=3.1.2
itsdangerous>=2.2
blinker>=1.9
click>=8.1.0
typing-extensions>=4.6.0
greenlet>=1
cffi>=1.12
pycparser
MarkupSafe>=2.0
requests>=2.32.3
python-dotenv>=1.0.0
PyPDF2>=3.0.1
python-docx>=0.8.11
anthropic>=0.3.11
google-generativeai>=0.3.2
PyJWT>=2.8.0
openai>=1.0.0
dashscope>=1.14.0
urllib3>=1.26.0,<2.0.0
